feat(admin): Teachers can only see their players
