+++ packages/server/prisma/schema.prisma
@@ -34,7 +34,7 @@ model Players {
   gameId          Int
   team            Team            @relation(fields: [teamId], references: [id])
   teamId          Int
-  user            User            @relation(fields: [userId], references: [id])
+  user            User            @relation("PlayedGames", fields: [userId], references: [id])
   userId          Int
   profile         Profile?        @relation(fields: [profileId], references: [id])
   profileId       Int?            @unique
@@ -50,7 +50,7 @@ model User {
   email       String    @unique
   firstName   String
   lastName    String
-  playedGames Players[]
+  playedGames Players[] @relation("PlayedGames")
   taughtGames Game[]    @relation("TaughtGames")
   roleId      Int
   role        Role      @relation(fields: [roleId], references: [id])
